Figured what was wrong
OK, so the problem was that the solution wasn't serving static files and that's why i needed to do a ng serve and run the app at the same time.
I had to actually manually create wwwroot folder (ignored in git...) and then ng build for the solution to get the wwwroot folder. Also, VS got stupid and needed to get restarted. After that, the server served the static files from wwwroot like angular was working.
Also made some small changes to startup and launch settings.