Switched from namespaces to modules
OK, so I don't know since when, but typescript isn't about namespaces
anymore. It's now about modules. And every ts file is a module. Which
kinda sucks. So I had to rewrite my ts models and interfaces. But worry
not, it's handled. There are now some .ts files which serve as
catalogues.